Rewrite bug search for clarity and performance
Having Bug sit occasionally on the bug_with_details view, with
magical (and multiple) seralizers was bad for the brain, and
ultimately led to some fairly inevitable N+1 query issues.

This commit replaces that view with a bug_summaries view tailored to
the API's search/show functionality, and a simple BugSummary model
which sits atop it. This makes the code much more comprehensible and
magic, and as a happy side effect, all N+1 queries have also been
eliminated, and the underlying SQL made more efficient.